+++ src/components/Add-New/addNew.jsx
@@ -1,5 +1,6 @@
 import { useSelector } from "react-redux";
 import { Link } from "react-router-dom";
+import "./add-new.scss";
 
 function AddNew() {
   const currentUser = useSelector((state) => state.todos.AllUserInfo);
