+++ src/components/LightDarkMode/light-dark-mode.scss
@@ -1,0 +1,67 @@
+@import "../../colors.scss";
+@import "../../mixins.scss";
+
+.main-container-relative {
+  width: 70px;
+  height: 40px;
+  position: relative;
+  display: flex;
+  justify-content: center;
+  padding: 0;
+  margin: 0;
+  background-color: inherit;
+  .wrapper {
+    width: 100%;
+    height: 80%;
+    position: absolute;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    top: 8%;
+    left: 8%;
+    padding: 0;
+    margin: 0;
+    background-color: inherit;
+    .design {
+      border: 1px solid var(--white, $white);
+      width: 100%;
+      height: 100%;
+      border-radius: 20px;
+      background-color: var(--lightWhite, $lightWhite);
+      z-index: 10;
+      position: relative;
+      font-size: 1.2rem;
+      font-family: sans-serif;
+      font-weight: bold;
+      text-align: center;
+      color: var(--colorLB);
+      display: flex;
+      align-items: center;
+      justify-content: center;
+      .togglebtn {
+        width: 25px;
+        height: 25px;
+        position: absolute;
+        display: flex;
+        align-items: center;
+        justify-content: center;
+        flex-flow: row wrap;
+        left: 5%;
+        border-radius: 50%;
+        background-color: var(--white, $white);
+        border: 1px solid var(--white, $white);
+        transition: left 1s;
+        cursor: pointer;
+      }
+      .fa-solid {
+        font-size: 1.2rem;
+      }
+      .day {
+        color: $yellow;
+      }
+      .night {
+        color: $green;
+      }
+    }
+  }
+}
