+++ src/components/Make-Todo/make-todo.scss
@@ -1,3 +1,6 @@
+@import "../../colors.scss";
+@import "../../mixins.scss";
+
 .todo-center {
   max-width: 1200px;
   padding: 1rem;
