+++ src/components/Modal/modal.scss
@@ -1,3 +1,6 @@
+@import "../../colors.scss";
+@import "../../mixins.scss";
+
 .modal {
   @include flex(row nowrap);
   background-color: $transparentBlack;
