+++ src/components/Search-Todo/search-todo.scss
@@ -1,0 +1,54 @@
+@import "../../colors.scss";
+@import "../../mixins.scss";
+
+.search-area {
+  @include flex(row nowrap);
+  padding: 0.5rem 0;
+  input {
+    width: 0%;
+    font-size: 2rem;
+    border-radius: 10px;
+    border: 1px solid rgba(34, 165, 150, 0.6);
+    opacity: 0;
+    transition: opacity 2s, width 2s;
+  }
+  div {
+    padding: 0 0 0 calc((100% - 60px) / 2); //center align
+    margin: auto;
+    transition: padding 1.9s; //To center align while transitioning
+  }
+  button {
+    background: $blue url("./images/search.png");
+    background-position: center;
+    background-size: 35px 35px;
+    background-repeat: no-repeat;
+    border: none;
+    padding: 0.5rem;
+    border-radius: 50%;
+    opacity: 1;
+    position: relative;
+    transition: opacity 1s;
+    width: 50px;
+    height: 50px;
+  }
+  button:hover {
+    background-color: rgb(1, 1, 191);
+    cursor: pointer;
+  }
+  button:active {
+    background-color: transparent;
+  }
+}
+//Animate searchBar
+.search-area:hover {
+  div {
+    padding: 0 0 0 calc((50% - 60px) / 2); //center align while input transition to different width
+  }
+  input {
+    opacity: 1;
+    width: 80%;
+  }
+  button {
+    opacity: 0;
+  }
+}
