+++ src/components/User-Welcome/user-welcome.scss
@@ -1,3 +1,6 @@
+@import "../../colors.scss";
+@import "../../mixins.scss";
+
 .welcome {
   display: flex;
   flex-flow: column nowrap;
