+++ src/pages/CreateNewTodo/_create-new-todo.scss
@@ -1,0 +1,60 @@
+.create-todos {
+  @include flex(column nowrap);
+  width: 100vw;
+  height: 100vh;
+  label {
+    width: 100%;
+    height: 90%;
+    max-width: 1200px;
+    textarea {
+      resize: none;
+      padding: 0.5rem;
+      width: 100%;
+      height: 100%;
+      background: $notePadColor;
+      color: var(--black, $black);
+      overflow-y: auto;
+      font-size: 1.2rem;
+    }
+  }
+  div {
+    width: 100%;
+    height: 10%;
+    background: $blue;
+    display: flex;
+    flex-flow: row nowrap;
+    justify-content: center;
+    div {
+      width: 100%;
+      max-width: 1200px;
+      height: 100%;
+      display: flex;
+      flex-flow: row nowrap;
+      justify-content: space-between;
+      align-items: center;
+      padding: 0 1rem;
+      img {
+        width: 60%;
+        height: 70%;
+        object-fit: contain;
+      }
+      button {
+        @include button-click-effect;
+        border: none;
+        background: transparent;
+        transition: transform 1s;
+        .fa-solid {
+          color: $white;
+          font-weight: bold;
+          font-size: 2rem;
+        }
+      }
+      button:hover {
+        transform: scale(1.1);
+      }
+      button:active {
+        transform: scale(0.5);
+      }
+    }
+  }
+}
