+++ src/pages/FogottenPasswordPage/ForgottenPassword.jsx
@@ -3,7 +3,7 @@ import { useEffect } from "react";
 import { useState } from "react";
 import { Navigate, useLocation } from "react-router-dom";
 import animation from "../../animations/popupbox";
-import LoadingScreen from "../../components/loadingScreen";
+import LoadingScreen from "../../components/Loading/loadingScreen";
 import { apiURL } from "../../store/actions";
 
 function ForgottenPassword() {
