+++ src/pages/UpdatePages/UpdateUsername.jsx
@@ -3,7 +3,7 @@ import { useState } from "react";
 import { useDispatch, useSelector } from "react-redux";
 import { Link, Navigate } from "react-router-dom";
 import animation from "../../animations/popupbox";
-import LoadingScreen from "../../components/loadingScreen";
+import LoadingScreen from "../../components/Loading/loadingScreen";
 import { apiURL } from "../../store/actions";
 import { todolistSlice } from "../../store/todolistSlice";
